# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Meno API
  version: 1.0.0
  description: >-
    RESTful API built with Express, MongoDB, and clean architecture principles. This API provides endpoints for
    authentication (email/password and Google OAuth), user management, and file storage with support for multiple
    storage providers (local and Google Cloud Storage).
  contact:
    name: API Support
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:6001
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authorization header using the Bearer scheme
...
paths:
...
  /api/projects/{projectId}/meetings:
    post:
      summary: Create a new meeting with audio file
      description: Upload audio file and create a meeting record
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
          description: Project ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audioFile
                - title
              properties:
                audioFile:
                  type: string
                  format: binary
                  description: Audio file (MP3, WAV, M4A, WebM, OGG - max 100MB)
                title:
                  type: string
                  minLength: 2
                  maxLength: 200
                  example: Weekly Team Standup
                recordingType:
                  type: string
                  enum:
                    - upload
                    - direct
                  default: upload
                duration:
                  type: number
                  format: float
                  description: Audio duration in seconds (optional, calculated from file if not provided)
                  example: 125.5
      responses:
        '201':
          description: Meeting created successfully
        '400':
          description: Validation error or invalid file
        '401':
          description: Unauthorized
        '404':
          description: Project not found
    get:
      summary: Get meetings in a project
      description: Retrieve paginated list of meetings for a project
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
          description: Project ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: sort
          schema:
            type: string
            default: '-createdAt'
      responses:
        '200':
          description: Meetings retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Project not found
  /api/projects/{projectId}/meetings/{id}:
    get:
      summary: Get meeting by ID
      description: Retrieve specific meeting details
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '200':
          description: Meeting retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
    put:
      summary: Update meeting
      description: Update meeting title
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 2
                  maxLength: 200
      responses:
        '200':
          description: Meeting updated successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
    delete:
      summary: Delete meeting
      description: Delete meeting, audio file, and all transcriptions
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '200':
          description: Meeting deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
  /api/projects/{projectId}/meetings/{id}/transcribe:
    post:
      summary: Start transcription for meeting
      description: Begin async transcription process for meeting audio
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '202':
          description: Transcription started successfully
        '400':
          description: Meeting already transcribed or in progress
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
  /api/projects/{projectId}/meetings/{id}/download:
    get:
      summary: Download meeting audio file
      description: Download the original audio file for a meeting
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '200':
          description: Audio file downloaded successfully
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            audio/mp4:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
  /api/projects/{projectId}/meetings/{id}/status:
    get:
      summary: Get transcription status
      description: Check transcription progress and status
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '200':
          description: Status retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
  /api/projects/{projectId}/meetings/{id}/summary/stream:
    post:
      summary: Generate meeting summary with streaming
      description: |
        Stream AI-generated meeting summary with real-time updates.
        Summary includes Overview, Key Points, Conclusion, and Action Items sections.

        IMPORTANT REQUIREMENTS:
        - Summary is in MARKDOWN FORMAT with proper headings (##) and bullet points (-)
        - Summary is generated in the SAME LANGUAGE as the meeting transcription
        - For example, if transcription is in Chinese, the summary will be in Chinese

        Saves to database when complete.
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
          description: Project ID
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '200':
          description: Streaming response with Server-Sent Events
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE stream with JSON events:
                  - {"type":"connected","meetingId":"..."}
                  - {"type":"chunk","content":"markdown text"}
                  - {"type":"complete"}
                  - {"type":"error","message":"..."}

                  Summary is in MARKDOWN FORMAT with the following sections:

                  ## Overview
                  Brief overview of meeting discussion

                  ## Key Points
                  - Main topics and decisions

                  ## Conclusion
                  Final outcomes and agreements

                  ## Action Items
                  - Tasks with responsibilities

                  Notes:
                  - All content is in MARKDOWN FORMAT (## for headings, - for bullet points)
                  - Summary is generated in the same language as the transcription
        '400':
          description: Transcription not completed
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
  /api/projects/{projectId}/meetings/{id}/action-items/generate:
    post:
      summary: Generate action items for meeting
      description: |
        Generate action items from meeting transcription using AI.
        Returns immediately with 202 status while processing in background.
        Client should poll meeting status to check completion.

        Action items include:
        - task: What needs to be done
        - assignee: Person responsible (matched from transcript speakers)
        - priority: high/medium/low
        - dueDate: When it should be completed
        - context: Additional context from the meeting
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
          description: Project ID
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '202':
          description: Action items generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Action items generation started
                  actionItemsStatus:
                    type: string
                    enum:
                      - processing
                    example: processing
                  actionItemsProgress:
                    type: number
                    example: 0
        '400':
          description: Transcription not completed or already processing
        '401':
          description: Unauthorized
        '404':
          description: Meeting not found
...
  /api/users/{userId}/meetings:
    get:
      summary: Get user's recent meetings
      description: Retrieve paginated list of meetings for a user across all their projects
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: User ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 5
        - in: query
          name: sort
          schema:
            type: string
            default: '-createdAt'
      responses:
        '200':
          description: User meetings retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: User not found
...